WIP: benchmarks/firecracker-boot: Initial implementation
This is an attempt to do things manually. It turns out getting NixOS to
boot in this environment is a bit fiddly, I stopped when I realised it's
annoying to get it to work on a read-only FS image, but there's already
a project called microvm.nix that solved this, so I'll try that instead.

Checking this in anyway so I have a reference to the old code that
semi-worked.